requirements.txt
tensorflow
pillow
python 3.5